feat: add commitlint, prettier, build step, GitHub workflow
This turned out bigger than I intended. It started out with me wanting
to lint commits, but then kind of exploded.
