+++ .prettierrc.json
@@ -1,0 +1,5 @@
+{
+  "semi": false,
+  "arrowParens": "always",
+  "singleQuote": true
+}
